update timer to avoid overflow,  update rtos testcase to fix 32bit timer value

+++ tests/FreeRTOS/port.c
@@ -174,10 +174,10 @@ static void prvSetNextTimerInterrupt(void)
 	__asm volatile("add t0,t0,%0" :: "r"(configTICK_CLOCK_HZ / configTICK_RATE_HZ));
 	__asm volatile("csrw mtimecmp,t0");
 #endif
-	volatile uint32_t timer_value;
+	volatile uint64_t timer_value;
 	timer_value = TIMER;
 // 	timer_value += configTICK_CLOCK_HZ / configTICK_RATE_HZ;
-	timer_value += 10000 - 620;    // timer set to 100 Hz and corrected
+	timer_value += (uint64_t)(10000 - 620);    // timer set to 100 Hz and corrected
 	TIMER_CMP = timer_value;
   // printf("ISR\n");
 }
@@ -196,13 +196,13 @@ void vPortSetupTimer(void)
 	__asm volatile("csrs mie,%0"::"r"(0x80));
 #endif
 
-        volatile uint32_t timer_value;
+        volatile uint64_t timer_value;
         
         register_timer_isr();
         
         timer_value = TIMER;
         // timer_value += configTICK_CLOCK_HZ / configTICK_RATE_HZ;
-	timer_value += 10000 - 620;        // timer set to 100 Hz and corrected
+        timer_value += (uint64_t)(10000 - 620);        // timer set to 100 Hz and corrected
         TIMER_CMP = timer_value;
 }
 /*-----------------------------------------------------------*/
